+++ lib/Braintree/CustomerGateway.php
@@ -202,6 +202,12 @@ class CustomerGateway
                 ['paypal' => [
                     'payee_email',
                     'payeeEmail',
+                    'order_id',
+                    'orderId',
+                    'custom_field',
+                    'customField',
+                    'description',
+                    'amount',
                     ['shipping' =>
                         [
                             'firstName', 'lastName', 'company', 'countryName',