+++ lib/Braintree/Transaction.php
@@ -172,6 +172,7 @@ namespace Braintree;
  * @property-read Braintree\Disbursement $disbursementDetails populated when transaction is disbursed
  * @property-read Braintree\Dispute $disputes populated when transaction is disputed
  * @property-read Braintree\AuthorizationAdjustment $authorizationAdjustments populated when a transaction has authorization adjustments created when submitted for settlement
+ * @property-read \DateTime $authorizationExpiresAt DateTime authorization will expire
  *
  */
 
@@ -225,6 +226,7 @@ class Transaction extends Base
     // Industry Types
     const LODGING_INDUSTRY           = 'lodging';
     const TRAVEL_AND_CRUISE_INDUSTRY = 'travel_cruise';
+    const TRAVEL_AND_FLIGHT_INDUSTRY = 'travel_flight';
 
     /**
      * sets instance properties from an array of values