+++ tests/integration/CustomerTest.php
@@ -647,7 +647,7 @@ class CustomerTest extends Setup
         $this->assertTrue($result->success);
     }
 
-    public function testCreate_worksWithOrderPaymentPayPalNonceWithPayeeEmailSnakeCase()
+    public function testCreate_worksWithOrderPaymentPayPalNonceWithSnakeCase()
     {
         $http = new HttpClientApi(Braintree\Configuration::$global);
         $nonce = $http->nonceForPayPalAccount([
@@ -674,7 +674,7 @@ class CustomerTest extends Setup
         $this->assertTrue($result->success);
     }
 
-    public function testCreate_worksWithOrderPaymentPayPalNonceWithPayeeEmailCamelCase()
+    public function testCreate_worksWithOrderPaymentPayPalNonceWithCamelCase()
     {
         $http = new HttpClientApi(Braintree\Configuration::$global);
         $nonce = $http->nonceForPayPalAccount([
@@ -1175,7 +1175,7 @@ class CustomerTest extends Setup
         $this->assertEquals($result->customer->defaultPaymentMethod()->token, $paypalAccountToken);
     }
 
-    public function testUpdate_worksWithOrderPaymentPayPalNonceWithPayeeEmailSnakeCase()
+    public function testUpdate_worksWithOrderPaymentPayPalNonceWithSnakeCase()
     {
         $customerResult = Braintree\Customer::create([
             'creditCard' => [
@@ -1205,6 +1205,10 @@ class CustomerTest extends Setup
             'options' => [
                 'paypal' => [
                     'payee_email' => 'payee@example.com',
+                    'order_id' => 'merchant-order-id',
+                    'custom_field' => 'custom merchant field',
+                    'description' => 'merchant description',
+                    'amount' => '1.23',
                 ],
             ],
         ]);
@@ -1213,7 +1217,7 @@ class CustomerTest extends Setup
         $this->assertEquals($result->customer->defaultPaymentMethod()->token, $paypalAccountToken);
     }
 
-    public function testUpdate_worksWithOrderPaymentPayPalNonceWithPayeeEmailCamelCase()
+    public function testUpdate_worksWithOrderPaymentPayPalNonceWithCamelCase()
     {
         $customerResult = Braintree\Customer::create([
             'creditCard' => [
@@ -1243,6 +1247,10 @@ class CustomerTest extends Setup
             'options' => [
                 'paypal' => [
                     'payeeEmail' => 'payee@example.com',
+                    'orderId' => 'merchant-order-id',
+                    'customField' => 'custom merchant field',
+                    'description' => 'merchant description',
+                    'amount' => '1.23',
                     'shipping' => [
                         'firstName' => 'Andrew',
                         'lastName' => 'Mason',