+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.37.0
+* Add `processorResponseType` to `Transaction`, `AuthorizationAdjustment`, and `CreditCardVerification`.
+* Add `authorizationExpiresAt` to `Transaction`
+* Add support for additional PayPal options when vaulting a PayPal Order during customer update
+* Allow PayPal payment ID and payer ID to be passed during transaction create
+* Add `travel_flight` support to industry-specific data
+
 ## 3.36.0
 * Fix dispute results in transactions not showing the correct status sometimes
 * Add `externalVault` option to transaction sale